simplify per comments
